Ref: Mejoras a los parsers
* Responden a una interfaz comun ahora
* Saque mucho codigo repetido de los try catch
* Quedaron mucho mas sencillos de entender
* Se inicializan unos valores en el Sprite.hpp que estaban cuasando problemas

+++ src/Utils/Utils.hpp
@@ -49,8 +49,8 @@ typedef struct enemigo{
 }enemigo_t;
 
 typedef struct bloque{
-    unsigned short posX;
-    unsigned short posY;
+    int posX;
+    int posY;
 	uint8_t numeroRecorteX;
 	uint8_t numeroRecorteY;
 }bloque_t;
